common: Moved the USB initialisation code into usb.c
